+++ code/maven-web-demo/src/main/java/com/yhocn/kaoqin/controller/KaoQinController.java
@@ -28,6 +28,17 @@ public class KaoQinController {
     @RequestMapping("/getList1")
     public ModelAndView getList1(ModelAndView mv, KaoQin kaoQin, String c, String s_name, Integer page, HttpServletRequest request, HttpSession session) {
         LoginController e = new LoginController();
+        boolean pd=false;
+        for (int i=0;i<e.quanxian.size();i++){
+            if(e.quanxian.get(i).getView_name().equals("考勤表")&&e.quanxian.get(i).getSel().equals("√")){
+                pd=true;
+            }
+        }
+        if(!pd){
+            mv.addObject("msg", "无权限");
+            mv.setViewName("/main.jsp");
+            return mv;
+        }
         if (request.getParameter("s_name") == null) {
             s_name = "";
         }else{
@@ -125,6 +136,17 @@ public class KaoQinController {
     @RequestMapping(value = "/add")
     public ModelAndView add(ModelAndView mv, KaoQin kaoQin, String c) {
         LoginController e = new LoginController();
+        boolean pd=false;
+        for (int i=0;i<e.quanxian.size();i++){
+            if(e.quanxian.get(i).getView_name().equals("考勤表")&&e.quanxian.get(i).getAdd().equals("√")){
+                pd=true;
+            }
+        }
+        if(!pd){
+            mv.addObject("msg", "无权限");
+            mv.setViewName("/main.jsp");
+            return mv;
+        }
         c = e.a;
         int i = service.add(kaoQin, c);
         if (i > 0) {
@@ -150,6 +172,17 @@ public class KaoQinController {
     @RequestMapping("/update")
     public ModelAndView update(ModelAndView mv, KaoQin kaoQin, String c) {
         LoginController e = new LoginController();
+        boolean pd=false;
+        for (int i=0;i<e.quanxian.size();i++){
+            if(e.quanxian.get(i).getView_name().equals("考勤表")&&e.quanxian.get(i).getUpd().equals("√")){
+                pd=true;
+            }
+        }
+        if(!pd){
+            mv.addObject("msg", "无权限");
+            mv.setViewName("/main.jsp");
+            return mv;
+        }
         c = e.a;
         int i = service.update(kaoQin, c);
         if (i > 0) {
@@ -165,6 +198,17 @@ public class KaoQinController {
     @RequestMapping("/delete")
     public ModelAndView delete(ModelAndView mv, KaoQin kaoQin, String c) {
         LoginController e = new LoginController();
+        boolean pd=false;
+        for (int i=0;i<e.quanxian.size();i++){
+            if(e.quanxian.get(i).getView_name().equals("考勤表")&&e.quanxian.get(i).getSel().equals("√")){
+                pd=true;
+            }
+        }
+        if(!pd){
+            mv.addObject("msg", "无权限");
+            mv.setViewName("/main.jsp");
+            return mv;
+        }
         c = e.a;
         int i = service.delete(kaoQin, c);
         if (i > 0) {
